refactor the components Fixes #72
