+++ src/libs/ui/organism/index.ts
@@ -1,1 +1,1 @@
-export { default as Layout } from './layout/layout';
+
