+++ src/libs/ui/template/index.ts
@@ -1,0 +1,1 @@
+export { default as Layout } from './layout/layout';
